+++ CHANGELOG.md
@@ -27,6 +27,8 @@ CHANGELOG
   become `gmail.com`.
 * Additional `gmail.com` typos are now normalized when `hashEmail` is used.
   For example, `gmali.com` will become `gmail.com`.
+* When `hashEmail` is used, the local part of an email address is now
+  normalized to NFC.
 
 2.0.0 (2023-12-04)
 ------------------